+++ actions/upload-images-to-jfrog/src/github-action.ts
@@ -1,45 +1,17 @@
-import { exec } from '@actions/exec';
-import { setFailed, error, warning } from '@actions/core';
-import { getInputs } from './util/github/get-inputs';
+import { setFailed, getInput } from '@actions/core';
+import { AppsInput } from '@src/types';
+import { uploadImages } from './lib';
 
 export async function run(): Promise<void> {
   try {
-    const { apps, remoteTag, localTag, registryUrl, actionIfMissing } = getInputs();
+    const apps: AppsInput[] = JSON.parse(getInput('apps'));
+    const remoteTag = getInput('remoteTag');
+    const localTag = getInput('localTag');
+    const registryUrl = getInput('registryUrl');
+    const actionIfMissing = getInput('actionIfMissing');
 
-    for (const app of apps) {
-      const name = app.name
-      const imageName = `${name}:${localTag}`
+    await uploadImages({ apps, localTag, remoteTag, registryUrl, actionIfMissing });
 
-      let imageId = "";
-
-      const options = {
-          listeners: {
-              stdout: (data: any) => {
-                  imageId += data.toString();
-              }
-          }
-      };
-
-      await exec("docker", ["images", "-q", imageName], options)
-
-      if (imageId !== "") {
-          await exec("docker", ["tag", imageName, `${registryUrl}/${name}:${remoteTag}`]);
-          await exec("docker", ["push", `${registryUrl}/${name}:${remoteTag}`]);
-      } else {
-          switch(actionIfMissing) {
-              case "warning":
-                // code block
-                warning(`${app.name} image is not created`)
-                break;
-              case "error":
-                // code block
-                error(`${app.name} image is not created`)
-                break;
-              default:
-                // code block
-            }
-      }
-  }
   } catch (error) {
     setFailed(error as Error);
   }
