+++ actions/upload-images-to-jfrog/src/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './upload-images';
